Cleaning up code for secure api gateway demo
